Simplify datachannel plugin references
